+++ tt_metal/sources.cmake
@@ -31,6 +31,7 @@ set(TT_METAL_PUBLIC_API
     api/tt-metalium/experimental/device.hpp
     api/tt-metalium/experimental/disaggregation/kv_chunk_address_table.hpp
     api/tt-metalium/experimental/dispatch_context.hpp
+    api/tt-metalium/experimental/fabric/cluster.hpp
     api/tt-metalium/experimental/fabric/control_plane.hpp
     api/tt-metalium/experimental/fabric/edm_fabric_counters.hpp
     api/tt-metalium/experimental/fabric/fabric.hpp
